[fix](nereids) fix partition dest overwrite bug when cte as bc right
